pyproject.toml
[build-system]
requires = [ "setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gw-cli"
version = "2.0.0"
description = "Google Workspace CLI - Gmail, Docs, Sheets, Slides, Calendar, Forms & Drive from the command line"
readme = "README.md"
keywords = [ "google", "workspace", "cli", "gmail", "drive", "docs", "sheets"]
requires-python = ">=3.10"
dependencies = [
 "google-api-python-client>=2.168.0",
 "google-auth-httplib2>=0.2.0",
 "google-auth-oauthlib>=1.2.2",
 "keyring>=25.0.0",
]

[[project.authors]]
name = "Ying Cong"

[project.license]
text = "MIT"

[project.scripts]
gw = "gw.__main__:main"

[project.optional-dependencies]
test = [
    "pytest>=8.3.0",
]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.12.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["gw*"]
exclude = ["tests*", "docs*", "build", "dist"]
